Load data to Recyclers

+++ src/vendor_app/app/src/main/kotlin/cz/quanti/android/vendor_app/repository/purchase/dto/api/V2InvoiceApiEntity.kt
@@ -6,10 +6,10 @@ data class V2InvoiceApiEntity (
 )
 
 data class InvoiceApiEntity(
-    var id: Int = 0,
+    var id: Long = 0,
     var projectId: Int = 0,
     var quantity: Int = 0,
     var date: String = "",
-    var value: Long = 0,
+    var value: Double = 0.0,
     var currency: String = ""
 )
